FIX displayProject error
